pyproject.toml
[tool.poetry]
name = "jackson"
version = "0.0.0"
description = ""
authors = ["Lev Vereshchagin <mail@vrslev.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.4.0"
pydantic = {extras = ["dotenv"], version = "^1.9.0"}
JACK-Client = "^0.5.3"
PyYAML = "^6.0"
asyncer = "^0.0.1"

[tool.poetry.scripts]
jackson = "jackson.main:app"

[tool.poetry.group.dev.dependencies]
black = "^21.12-beta.0"
pre-commit = "^2.16.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
typeCheckingMode = "strict"
